+++ src/asset.rs
@@ -61,6 +61,7 @@ impl Assets {
             IconId::ScissorRectangle => &self.icons.scissor_rectangle,
             IconId::ScissorLasso => &self.icons.scissor_lasso,
             IconId::Lasso => &self.icons.lasso,
+            IconId::SelectBucket => &self.icons.select_bucket,
             IconId::GoLeft => &self.icons.go_left,
             IconId::GoRight => &self.icons.go_right,
             IconId::GoTop => &self.icons.go_top,
@@ -113,6 +114,7 @@ pub enum IconId {
     ScissorRectangle,
     ScissorLasso,
     Lasso,
+    SelectBucket,
     GoLeft,
     GoRight,
     GoTop,
@@ -281,6 +283,7 @@ pub struct Icons {
     pub scissor_rectangle: Sprite,
     pub scissor_lasso: Sprite,
     pub lasso: Sprite,
+    pub select_bucket: Sprite,
     pub go_left: Sprite,
     pub go_right: Sprite,
     pub go_top: Sprite,
@@ -333,6 +336,8 @@ impl Icons {
             scissor_lasso: decode_sprite(include_bytes!("../assets/icon-scissor-lasso.png"))
                 .or_fail()?,
             lasso: decode_sprite(include_bytes!("../assets/icon-lasso.png")).or_fail()?,
+            select_bucket: decode_sprite(include_bytes!("../assets/icon-select-bucket.png"))
+                .or_fail()?,
             go_left: go.clip(block).or_fail()?,
             go_right: go.clip(block.shift_x(1)).or_fail()?,
             go_top: go.clip(block.shift_x(2)).or_fail()?,