+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add bucket selecting tool
 - Add preview scale setting
 - Add import-image-from-clipboard feature
 - Add finger friendly drawing mode